Module Sc_report.Config

val section : Types.config Sc_config.Section.section

val css_file : Types.config -> string

Returns the css file content given the format in the configuration. Raises RESOURCE_ERROR Missing_resource_file if the CSS file is not present in the resource directory.