Module PM.TestRepr

include {REPR}1

module Params : Sc_lib.Ctypes_bridge.ENCODING_PARAMS

Encoding parameters for the representation

val size_bounds : Sc_lib.Ctypes_bridge.Struct.typ -> int * int

size_bounds typ returns a pair of the minimal structure size (i.e, assuming all pointers are NULL), and maximal size (i.e., the total memory footprint with pointers unfolded according to Encoding.max_ptr_array_length)

val fold_prims : ?root_access_path:Sc_lib.C_concept.rev_access_path -> f: (Sc_lib.C_concept.access_path -> 'a -> 'a) Sc_lib.Ctypes_bridge.ctyp_consumer -> Sc_lib.Ctypes_bridge.Struct.typ -> 'a -> 'a

fold_prims ~f typ folds f over every non-pointer field defined in type typ (cf Encoding.ctyp_fold_prims).

val fold_pointers : deep:bool -> ?root_access_path:Sc_lib.C_concept.rev_access_path -> ?max_depth:int -> ?unfold_implicit_arrays:bool -> f: (Sc_lib.C_concept.access_path -> Sc_lib.Ctypes_bridge.ptr_attributes -> 'a -> 'a) Sc_lib.Ctypes_bridge.ctyp_consumer -> Sc_lib.Ctypes_bridge.Struct.typ -> 'a -> 'a

fold_pointers ~deep ?unfold_implicit_arrays ~f typ folds f over every pointer field defined in type typ (cf Encoding.ctyp_fold_pointers); referenced types are visited recursively if deep holds.

module Val : sig ... end