Module Report_builder.MakeReportBuilder

Parameters

module C : ReportConfig

Signature

type tool_config = (string * string) list

A tool configuration is a list of pairs (key * value).

type run_config = tool_config accordion list

A run configuration is a list of tool configurations.

type run = {

run_time : float;
(*
The time of the run.
*)

run_config : run_config;
(*
The run configuration.
*)

extra_run_html : string;
(*
Some extra custom HTML.
*)

}

type run_tabs = run tab list

A list of tabs, each for one run.

type overview = {

num_tests_gen : int;

num_tests_imported : int;

num_labels : int;

num_covered : int;

num_uncoverable : int;

num_runs : int;

runs : run_tabs;

}

The overview page.

type test = {

tool_name : string;
(*
The tool that generated the test.
*)

tool_time : float;
(*
The time it took to generate the test.
*)

test_file : [ `C ] Sc_sys.File.TYPES.file;
(*
The test file name. Will be opened.
*)

}

A single test.

type tests = test tab list

A list of tests.

type label = {

static_db_data : Sc_lib.Types.Label.simple_view;

runtime_db_data : [ `Cov | `Uncov | `Unk ];

}

All the information on a label from the static and runtime databases.

type report_tab =

| Overview of overview

| Tests of tests

| Labels of label tab list

Each possible tab.

type report = {

r_title : string;

r_tabs : report_tab tab list;

}

module Report : ToHtml with type t = report